requirements.txt
docling
gradio
--extra-index-url https://download.pytorch.org/whl/cu118
torch
pytesseract
python-docx
markdown
requests